adds Type events
