+++ list/list-test.js
@@ -1278,8 +1278,8 @@ QUnit.test("iterator can recover from bad _length", function() {
 QUnit.test("onPatches", function(){
 	var list = new DefineList(["a","b"]);
 	var PATCHES = [
-		[ {deleteCount: 2, index: 0} ],
-		[ {index: 0, insert: ["A","B"], deleteCount: 0} ]
+		[ {deleteCount: 2, index: 0, type: "splice"} ],
+		[ {index: 0, insert: ["A","B"], deleteCount: 0, type: "splice"} ]
 	];
 	var handlerCalls = 0;
 	var handler = function(patches){
