+++ test/test-define-only.js
